Move coordinate generation funcs and block_reduce
Move these functions to new modules `coordinates.py` and `grid_math.py`
to keep `utils.py` more manageable.
